# openapi_1.yaml
openapi: 3.0.3
info:
  title: Music & Social
  version: 1.0.4
...
paths:
...
  /users:
    post:
      tags: 
        - auth
      description: Anlegen eines Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Users-request"
      responses:
        "201":
          description:
            User wurde angelegt
        "403":
          description:
            Forbidden
  /users/{userId}:
    get:
      tags:
        - auth
      description: Gibt einen User zurück
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            properties:
              id:
                type: number
      responses:
        "200":
          description:
            User wurde zurückgegeben
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description:
            Vorgang nicht möglich Authorisierung notwendig
        "404":
          description:
            User nicht gefunden
        "410":
          description:
            Resource steht nicht mehr zur verfügung
    patch:
      tags:
        - auth
      summary: Updaten eines Users
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            properties:
              id:
                type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User wurde erfolgreich geändert
        '400': 
          description: User konnte nicht geändert werden
        '404':
          description: User existiert nicht
        "403":
          description:
            Forbidden
        "410":
          description:
            Resource steht nicht mehr zur verfügung
    delete:
      tags:
        - auth
      summary: Löschen eines Nutzers
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            properties:
              id:
                type: integer
      responses:
        '200':
          description: User wurde erfolgreich gelöscht
        '400':
          description: User konnte nicht gelöscht werden
        "401":
          description:
            Vorgang nicht möglich Authorisierung notwendig
        "404":
          description:
            User existiert nicht
        "410":
          description:
            Resource steht nicht mehr zur Verfügung
...
  /posts:
    post:
      tags:
        - posts
      summary: Creates new post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Post'
      responses:
        default:
          description: successful operation
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
  /posts/{postId}:
    get:
      tags:
        - posts
      summary: Return post by ID
      parameters:
        - name: postId
          required: true
          in: path
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: operation was unsuccessful
        "401":
          description:
            Vorgang nicht möglich Authorisierung notwendig
        "410":
          description:
            Resource steht nicht mehr zur verfügung
    patch:
      tags:
        - posts
      summary: Aktualisieren eines Posts
      parameters:
        - name: postId
          required: true
          in: path
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Post'
      responses:
        '200':
          description: Post wurde erfolgreich aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Post konnte nicht aktualisiert werden
        '404': 
          description: Post existiert nicht
        "401":
          description:
            Vorgang nicht möglich Authorisierung notwendig
    delete:
      tags:
        - posts
      summary: Entfernen eines Posts
      parameters:
        - name: postId
          required: true
          in: path
          schema:
            type: integer
      responses:
        '200': 
          description: Post wurde erfolgreich entfernt
        '400':
          description: Post konnte nicht entfernt werden
        '404': 
          description: Post existiert nicht
        "401":
          description:
            Vorgang nicht möglich Authorisierung notwendig
...
components:
  schemas:
...
    Users-request:
      type: object
      properties:
        id:
          type: number
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        role:
          type: string
        password:
          type: string
        createdAt:
          type: string
          format: date
    # -------------------- vllt nochmal überarbeiten --------------------
        auths: 
          type: array
          items:
            type: string
    # ------------------------------------------------------------------
    Post:
      type: object
      properties:
        id:
          type: integer
        list:
          type: number
        type:
          type: string
        createdAt:
          type: string
          format: date
        updatedAt:
          type: string
          format: date